requirements.txt
fastapi[all]
psycopg2-binary
sqlalchemy
pyjwt
python-dotenv